+++ internal/product/data.go
@@ -1,9 +1,9 @@
 package product
 
 type Product struct {
-	ID          uint   `json:"id"`
+	ID          int32  `json:"id"`
 	Title       string `json:"title"`
 	Description string `json:"description"`
-	Amount      uint64 `json:"amount"`
+	Amount      int32  `json:"amount"`
 	IsGift      bool   `json:"is_gift"`
 }
